Move subcommands and use the registration decorator
